Check `clone_or_copy_needed` in `unnecessary_filter_map::check`
